More CCubeMaterial, more CGraphics, more aurora::gx

+++ Runtime/Graphics/CCubeMaterial.hpp
@@ -96,7 +96,7 @@ private:
   static void SetupBlendMode(u32 blendFactors, const CModelFlags& flags, bool alphaTest);
   static void HandleDepth(CModelFlagsFlags modelFlags, CCubeMaterialFlags matFlags);
   static u32 HandleColorChannels(u32 chanCount, u32 firstChan);
-  static void HandleTev(u32 tevCur, const u8* materialDataCur, const u32* texMapTexCoordFlags, bool shadowMapsEnabled);
+  static void HandleTev(u32 tevCur, const u32* materialDataCur, const u32* texMapTexCoordFlags, bool shadowMapsEnabled);
   static u32 HandleAnimatedUV(const u32* uvAnim, u32 texMtx, u32 pttTexMtx);
   static void HandleTransparency(u32& finalTevCount, u32& finalKColorCount, const CModelFlags& modelFlags,
                                  u32 blendFactors, u32& finalCCFlags, u32& finalACFlags);
